Entity fixes
all colors are now output as .0 floats to resolve read issues in G3D.
FGD types of angle or angles will allow the render of direction in 2d views.
Extended render of angle line in 2d views.
Angles will now be read in correctly and written in 3dt format correctly.